refactor(api): reuse shared request security manager
