+++ apps/api/src/flows.e2e.spec.ts
@@ -7,7 +7,10 @@ import { ed25519 } from '@noble/curves/ed25519';
 import { bytesToHex, hexToBytes } from '@noble/curves/abstract/utils';
 import { AppModule } from './app.module.js';
 import { InMemoryBankService, getFallbackBankService } from './in-memory-bank.service.js';
-import { createSignaturePayload, serializeBody } from './request-security.js';
+import {
+  createSignaturePayloadFromComponents as createSignaturePayload,
+  serializeBody,
+} from '@qzd/shared/request-security';
 import { createMockCardKeys, createOfflineVoucher } from '@qzd/card-mock';
 
 const DEV_SIGNING_PRIVATE_KEY_HEX =
